Change of direction: Convert once; Offer only mp3 on FE
It seems like I will never get on-the-fly conversion to work properly. I gave up when I figured out that in order to
write the Lame/Xing header for seeking and correct duration ffmpeg needs to seek in the output file.

That makes sense but I don't think I'll find a good solution to that.

New direction: Convert once. Show status of conversion. Scan automatically.